Add Folder Logic
-Changes the default directory for movies to be the DCIM/Camera. This
allows the player to easily play movies recorded on Glass itself.

-Adds support for subfolders containing video files. Subfolders will
sorted to the top of the list and can be tapped to navigate into them. A
down-swipe will take you back one level.

-The filename will not be read aloud if it contains nothing but numbers
and underscores. This is because it is not useful to have it read the
timestamp names of Glass-recorded videos as a huge number.